feat(entra-id): make the Entra ID crawler self-contained via the wizard plugin system
Relocates the Microsoft Graph crawler's wizard (CrawlerMeta.js, ConfigWizard.jsx,
Summary.jsx, discover.js) into tools/crawlers/entra-id/, the last major built-in
crawler still hardcoded in CrawlersPage.jsx/jobs.js. Follows the same pattern
already used by csv, omada, and midpoint.

Backend: folds the two bespoke routes (validate-graph-credentials,
discover-graph-attributes) and their entra-id-only constants/helpers into a
single tools/crawlers/entra-id/discover.js, loaded by the existing generic
POST /api/admin/crawlers/:type/discover route.

Frontend: moves the 565-line EntraIdWizard + AttributePicker + constants into
ConfigWizard.jsx with the standard 5-prop contract and a self-saving handleSave
(matching csv/omada/midpoint), and the card summary block into Summary.jsx.
Deletes the now-dead 'entra-wizard' dispatch branches from CrawlersPage.jsx.

Fixes two real bugs found while relocating:
- signInLogsDays/aiNamePatterns (Advanced Options) were silently dropped on
  save because handleWizardComplete's destructuring never picked them up.
- Editing a crawler without re-entering the secret broke attribute discovery
  (400) because it read the stripped clientSecret from the stored config
  instead of the vault — now resolves it via getConfigSecret like every
  other crawler type.

Also fixes a PowerShell nightly integration test (Test-EntraIdCrawler.ps1)
that called the old /admin/validate-graph-credentials route directly.
