+++ tools/crawlers/entra-id/CrawlerMeta.js
@@ -1,0 +1,5 @@
+export default {
+  id: 'entra-id',
+  name: 'Microsoft Graph',
+  description: 'Sync users, groups, roles, and governance data from Entra ID',
+};
